refactor: split recipe resolution into fetch and render stages
`canister::recipe::handlebars` did two unrelated jobs in one method: it
fetched a recipe's Handlebars template (local read, HTTP download, package
cache) and then rendered that template into build/sync steps. Rendering was
therefore only reachable through something that owned an HTTP client and a
package cache, and could not be tested without writing a temp file first.

Split it in two:

- `recipe::fetch::RecipeFetcher` (was `Handlebars`) retrieves the template
  text and nothing else, behind the `Resolve` seam, which now returns a
  `String` instead of `(BuildSteps, SyncSteps)`.
- `recipe::render_recipe` is a pure function from template text to steps. Its
  five tests are plain `#[test]`s over string inputs; the four they replace
  each needed a tempdir, a package cache and a tokio runtime.

Two behavioural notes:

- A rendered document that is not a valid build/sync manifest was a `panic!`;
  it is now `RenderRecipeError::Parse`. The rendered YAML moves to a `debug!`
  line, so `--debug` still shows it.
- A remote template is now cached once its checksum verifies, rather than
  after it also renders. The checksum is what establishes the bytes are the
  bytes we asked for, and rendering depends on per-canister context that has
  nothing to do with the template's cacheability.

`ConsolidateManifestError::Recipe` splits into `FetchRecipe` and
`RenderRecipe` accordingly, so a failure names the stage that failed.

Salvaged from PR 660, originally authored by Adam Spofford, re-executed on
current `main`: that branch introduced this split as part of a larger crate
reorganization, and the split stands on its own.
