+++ crates/icp/src/canister/recipe/mod.rs
@@ -1,55 +1,36 @@
+//! Recipe resolution, split into two stages.
+//!
+//! [`fetch`] retrieves a recipe's Handlebars template — reading a local file, or
+//! downloading and caching a remote URL or registry recipe — and returns the raw
+//! template text. [`render`] turns that text into concrete build/sync steps. The
+//! first stage does I/O and nothing else; the second is a pure function.
+//!
+//! The [`Resolve`] seam therefore covers only the fetching half, so a caller that
+//! already has a template (or must not touch the network) can render without
+//! going through a resolver at all.
+
 use async_trait::async_trait;
 use snafu::prelude::*;
 
-use crate::manifest::{
-    canister::{BuildSteps, SyncSteps},
-    recipe::Recipe,
-};
+use crate::manifest::recipe::Recipe;
 
-pub mod handlebars;
+pub mod fetch;
+pub mod render;
 
-/// Context passed to a recipe resolver, describing the canister being built.
+pub use render::{RecipeContext, RenderRecipeError, render_recipe};
+
+/// Retrieves the recipe templates a project references.
 ///
-/// Serializes to the shape injected into recipe templates under the `_` namespace:
-///
-/// ```yaml
-/// canister:
-///   name: <canister_name>
-/// ```
-pub struct RecipeContext {
-    pub canister_name: String,
-}
-
-impl RecipeContext {
-    /// Builds the YAML value injected into recipe templates under the `_` namespace.
-    /// Constructing the mapping directly is infallible, unlike `serde` serialization.
-    pub fn to_yaml(&self) -> serde_yaml::Value {
-        use serde_yaml::{Mapping, Value};
-
-        let mut canister = Mapping::new();
-        canister.insert("name".into(), Value::String(self.canister_name.clone()));
-
-        let mut root = Mapping::new();
-        root.insert("canister".into(), Value::Mapping(canister));
-
-        Value::Mapping(root)
-    }
-}
-
-/// A recipe resolver takes a recipe that is specified in a canister manifest
-/// and resolves it into a set of build/sync steps
+/// Only *fetching* is behind this trait: rendering a fetched template into build
+/// and sync steps is [`render_recipe`], which needs no I/O and so needs no seam.
 #[async_trait]
 pub trait Resolve: Sync + Send {
-    #[allow(clippy::result_large_err)]
-    async fn resolve(
-        &self,
-        recipe: &Recipe,
-        recipe_context: &RecipeContext,
-    ) -> Result<(BuildSteps, SyncSteps), ResolveError>;
+    /// Fetch the Handlebars template for `recipe`, returning its raw source.
+    async fn resolve(&self, recipe: &Recipe) -> Result<String, ResolveError>;
 }
 
 #[derive(Debug, Snafu)]
 pub enum ResolveError {
-    #[snafu(display("failed to resolve handlebars template"))]
-    Handlebars { source: handlebars::HandlebarsError },
+    #[snafu(display("failed to fetch recipe template"))]
+    Fetch { source: fetch::RecipeFetchError },
 }
